Move deprecated vault to :data
